Issue #1405: FolderController moved to folder package
