+++ api/src/test/java/com/epam/pipeline/controller/folder/FolderControllerTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.epam.pipeline.controller.pipeline;
+package com.epam.pipeline.controller.folder;
 
 import com.epam.pipeline.entity.metadata.FolderWithMetadata;
 import com.epam.pipeline.entity.metadata.MetadataEntity;
